+++ ClimaNotificacoesAPI.Application/Services/UsuarioService.cs
@@ -1,3 +1,4 @@
+using ClimaNotificacoesAPI.Application.Exceptions;
 using ClimaNotificacoesAPI.Domain.Entities;
 using ClimaNotificacoesAPI.Domain.Interfaces;
 using Microsoft.AspNetCore.Identity;
@@ -22,17 +23,37 @@ public class UsuarioService
     }
     public async Task<Usuario> CreateAsync(Usuario usuario)
     {
+        var usuarioExistente = await GetByEmailAsync(usuario.Email);
+
+        if (usuarioExistente != null)
+            throw new EmailJaCadastradoException(usuario.Email);
 
         usuario.Senha = _passwordHasher.HashPassword(usuario, usuario.Senha);
         return await _usuarioRepository.AddAsync(usuario);
     }
     public async Task<Usuario> UpdateAsync(Usuario usuario)
     {
-        usuario.Senha = _passwordHasher.HashPassword(usuario, usuario.Senha);
-        return await _usuarioRepository.UpdateAsync(usuario);
+        var usuarioExistente = await GetByIdAsync(usuario.Id);
+        if (usuarioExistente == null)
+        {
+            throw new UsuarioNaoEncontradoException();
+        }
+        var usuarioComMesmoEmail = await GetByEmailAsync(usuario.Email);
+        if (usuarioComMesmoEmail != null && usuarioComMesmoEmail.Id != usuario.Id)
+            throw new EmailJaCadastradoException(usuario.Email);
+
+        usuarioExistente.Nome = usuario.Nome;
+        usuarioExistente.Email = usuario.Email;
+        usuarioExistente.Senha = _passwordHasher.HashPassword(usuario, usuario.Senha);
+        return await _usuarioRepository.UpdateAsync(usuarioExistente);
     }
     public async Task DeleteAsync(int id)
     {
+        var usuarioExistente = await GetByIdAsync(id);
+        if (usuarioExistente == null)
+        {
+            throw new UsuarioNaoEncontradoException();
+        }
         await _usuarioRepository.DeleteAsync(id);
     }
     public async Task<Usuario> GetByEmailAsync(string email)
@@ -41,6 +62,16 @@ public class UsuarioService
     }
     public async Task<IEnumerable<Cidade>> GetCidadesByUsuarioIdAsync(int usuarioId)
     {
-        return await _usuarioRepository.GetCidadesByUsuarioIdAsync(usuarioId);
+        var usuario = await GetByIdAsync(usuarioId);
+        if (usuario == null)
+        {
+            throw new UsuarioNaoEncontradoException();
+        }
+        var cidades = await _usuarioRepository.GetCidadesByUsuarioIdAsync(usuarioId);
+        if (cidades == null || !cidades.Any())
+        {
+            throw new CidadeNaoEncontradaException("Cidades");
+        }
+        return cidades;
     }
 }
